pyproject.toml
[project]
name = "intentc"
version = "0.1.0"
description = "A compiler of intent - transforms specs into working code using AI agents"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.scripts]
intentc = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/core", "src/parser", "src/config", "src/graph", "src/git", "src/agent", "src/builder", "src/cli", "src/validation", "src/state"]

[tool.pytest.ini_options]
testpaths = ["src"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[tool.hatch.build.targets.sdist]
include = ["src/"]
